requirements.txt
pip>=21.0.1
numpy
pandas
jsonschema>=3.2.0
pycodestyle>=2.6.0
flask>=1.1.2
flask-cors==1.10.3
flask-restx==0.5.0
Flask-RESTful>=0.3.8
mysqlclient>=2.1.0
apache-airflow>=1.10.10
requests==2.28.0
pymongo>=4.1.1
influxdb-client==1.29.1
airflow-code-editor>=5.2.2
black>=22.3.0
requests
minio
influxdb